+++ sdk/dataprotection/azure-mgmt-dataprotection/generated_samples/resource_guard_crud/delete_resource_guard.py
@@ -35,6 +35,6 @@ def main():
     )
 
 
-# x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-01-01/examples/ResourceGuardCRUD/DeleteResourceGuard.json
+# x-ms-original-file: specification/dataprotection/resource-manager/Microsoft.DataProtection/stable/2023-05-01/examples/ResourceGuardCRUD/DeleteResourceGuard.json
 if __name__ == "__main__":
     main()